net: add obj_cnt track base code
This patch is to create the base code for obj_cnt track, obj can be
dev, dst, in6_dev, xfrm_state etc. It builds a hashtable to collect
the data of obj_hold/put(), including the call trace and the counts
for each call for some objs, which can be set by sysctl.

3 APIs are exported and can be used for track one specific type of
obj:

  1. bool obj_cnt_allowed(net, index, name, obj):
     check if one obj with index/name/obj can be track.

  2. void obj_cnt_track(net, loc, obj, type);
     save count this obj and save its call trace, type is to identify
     the obj type and put/hold type.

  3. void obj_cnt_dump(net, obj, type, str);
     dump the obj track data with obj/str and type.

For examples how to use them, see the following patches.

Signed-off-by: Xin Long <lucien.xin@gmail.com>